font settings

+++ trunk/cpsycle/audio/src/luaplugin.c
@@ -22,9 +22,9 @@ static void seqtick(psy_audio_LuaPlugin*, uintptr_t channel,
 	const psy_audio_PatternEvent*);
 static void sequencerlinetick(psy_audio_LuaPlugin*);
 static psy_audio_MachineInfo* info(psy_audio_LuaPlugin*);
-static void parametertweak(psy_audio_LuaPlugin*, uintptr_t par, int val);
+static void parametertweak(psy_audio_LuaPlugin*, uintptr_t par, float val);
 static int describevalue(psy_audio_LuaPlugin*, char* txt, uintptr_t param, int value);
-static int parametervalue(psy_audio_LuaPlugin*, uintptr_t param);
+static float parametervalue(psy_audio_LuaPlugin*, uintptr_t param);
 static void dispose(psy_audio_LuaPlugin*);
 static int mode(psy_audio_LuaPlugin*);
 static uintptr_t numinputs(psy_audio_LuaPlugin*);
@@ -195,7 +195,7 @@ psy_audio_MachineInfo* info(psy_audio_LuaPlugin* self)
 	return self->plugininfo;
 }
 
-void parametertweak(psy_audio_LuaPlugin* self, uintptr_t par, int val)
+void parametertweak(psy_audio_LuaPlugin* self, uintptr_t par, float val)
 {
 
 }
@@ -206,9 +206,9 @@ int describevalue(psy_audio_LuaPlugin* self, char* txt, uintptr_t param,
 	return 0;
 }
 
-int parametervalue(psy_audio_LuaPlugin* self, uintptr_t param)
+float parametervalue(psy_audio_LuaPlugin* self, uintptr_t param)
 {
-	return 0;
+	return 0.f;
 }
 
 int mode(psy_audio_LuaPlugin* self)
