+++ trunk/cpsycle/host/src/workspace.h
@@ -140,7 +140,7 @@ void workspace_stopfollowsong(Workspace*);
 void workspace_idle(Workspace*);
 void workspace_showparameters(Workspace*, uintptr_t machineslot);
 void workspace_selectview(Workspace*, int view);
-void workspace_parametertweak(Workspace*, int slot, uintptr_t tweak, int value);
+void workspace_parametertweak(Workspace*, int slot, uintptr_t tweak, float value);
 void workspace_recordtweaks(Workspace*);
 void workspace_stoprecordtweaks(Workspace*);
 int workspace_recordingtweaks(Workspace*);
